requirements.txt
# Core Django packages
Django==5.2.4
django-allauth
django-crispy-forms
crispy-tailwind
psycopg2-binary
spacy
faiss-cpu
sentence-transformers
django-htmx
pgvector
djangorestframework
gunicorn
django-cors-headers
pymupdf

# Force CPU-only torch
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
